Fix: create Loading component to get Evaluations from API and fix error
...with dependencies on useEffect from App.
