+++ src/components/App.js
@@ -1,7 +1,6 @@
 import React, { useEffect } from 'react';
 import { BrowserRouter, Switch, Route } from 'react-router-dom';
 import { useSelector } from 'react-redux';
-import _ from 'lodash';
 import Home from './Home';
 import Login from './registrations/Login';
 import Signup from './registrations/Signup';
@@ -11,26 +10,17 @@ import '../assets/styles/App.scss';
 import EvaluationForm from './EvaluationForm';
 import EvaluationsList from './EvaluationsList';
 import MoreMenu from './MoreMenu';
-import useGetEvaluations from '../hooks/useGetEvaluations';
 import Error from './Error';
 import useLoginStatus from '../hooks/useLoginStatus';
+import Loading from './Loading';
 
 const App = () => {
-  const getEvaluations = useGetEvaluations();
-  const currentEvaluation = useSelector(state => state.evaluations.currentEvaluation);
-  const isLoggedIn = useSelector(state => state.auth.isLoggedIn);
   const error = useSelector(state => state.error);
   const loginStatus = useLoginStatus();
 
   useEffect(() => {
     loginStatus();
   });
-
-  useEffect(() => {
-    if (isLoggedIn && _.isEmpty(currentEvaluation)) {
-      getEvaluations();
-    }
-  }, [isLoggedIn]); // eslint-disable-line
 
   if (error) {
     return <Error />;
@@ -41,7 +31,8 @@ const App = () => {
       <BrowserRouter>
         <Header />
         <Switch>
-          <Route exact path="/" component={Home} />
+          <Route exact path="/" component={Loading} />
+          <Route path="/home" component={Home} />
           <Route path="/login" component={Login} />
           <Route path="/signup" component={Signup} />
           <Route path="/eval" component={EvaluationForm} />
