+++ src/reducers/evaluationsReducer.js
@@ -25,6 +25,7 @@ const evaluationsReducer = (state = initialState, action) => {
       };
     }
     case GET_EVALUATIONS: {
+      console.log('get eval reducer')
       const currentEvaluation = action.payload.filter(evaluation => (
         evaluation.created_at === moment().format('DD MMM YY')
       )).map(curEval => formatData(curEval));
